выделил из страницы обращения, компонент с подробной информацией предложения. предусмотрел пустое детальное предложение для корректного ожидания загрузки информации.
